# ZTS: Replace MD5 and SHA256 wit XXH128