--- a/.github/workflows/scripts/qemu-3-deps.sh
+++ b/.github/workflows/scripts/qemu-3-deps.sh
@@ -16,7 +16,7 @@ function archlinux() {
   sudo pacman -Sy --noconfirm base-devel bc cpio dhclient dkms fakeroot \
     fio gdb inetutils jq less linux linux-headers lsscsi nfs-utils parted \
     pax perf python-packaging python-setuptools qemu-guest-agent ksh samba \
-    sysstat rng-tools rsync wget
+    sysstat rng-tools rsync wget xxhash
   echo "##[endgroup]"
 }
 
@@ -38,7 +38,7 @@ function debian() {
     lsscsi nfs-kernel-server pamtester parted python3 python3-all-dev \
     python3-cffi python3-dev python3-distlib python3-packaging \
     python3-setuptools python3-sphinx qemu-guest-agent rng-tools rpm2cpio \
-    rsync samba sysstat uuid-dev watchdog wget xfslibs-dev zlib1g-dev
+    rsync samba sysstat uuid-dev watchdog wget xfslibs-dev  xxhash zlib1g-dev
   echo "##[endgroup]"
 }
 
@@ -48,8 +48,7 @@ function freebsd() {
   echo "##[group]Install Development Tools"
   sudo pkg install -y autoconf automake autotools base64 checkbashisms fio \
     gdb gettext gettext-runtime git gmake gsed jq ksh93 lcov libtool lscpu \
-    pkgconf python python3 pamtester pamtester qemu-guest-agent rsync \
-    sysutils/coreutils
+    pkgconf python python3 pamtester pamtester qemu-guest-agent rsync xxhash
   sudo pkg install -xy \
     '^samba4[[:digit:]]+$' \
     '^py3[[:digit:]]+-cffi$' \
@@ -76,7 +75,7 @@ function rhel() {
     lsscsi mdadm nfs-utils openssl-devel pam-devel pamtester parted perf \
     python3 python3-cffi python3-devel python3-packaging kernel-devel \
     python3-setuptools qemu-guest-agent rng-tools rpcgen rpm-build rsync \
-    samba sysstat systemd watchdog wget xfsprogs-devel zlib-devel
+    samba sysstat systemd watchdog wget xfsprogs-devel xxhash zlib-devel
   echo "##[endgroup]"
 }