--- a/tests/zfs-tests/include/commands.cfg
+++ b/tests/zfs-tests/include/commands.cfg
@@ -100,7 +100,8 @@ export SYSTEM_FILES_COMMON='awk
     uniq
     vmstat
     wc
-    xargs'
+    xargs
+    xxh128sum'
 
 export SYSTEM_FILES_FREEBSD='chflags
     compress
@@ -112,13 +113,11 @@ export SYSTEM_FILES_FREEBSD='chflags
     jexec
     jls
     lsextattr
-    md5
     mdconfig
     newfs
     pw
     rmextattr
     setextattr
-    sha256
     showmount
     swapctl
     sysctl
@@ -146,7 +145,6 @@ export SYSTEM_FILES_LINUX='attr
     lscpu
     lsmod
     lsscsi
-    md5sum
     mkswap
     modprobe
     mountpoint
@@ -156,7 +154,6 @@ export SYSTEM_FILES_LINUX='attr
     perf
     setfattr
     setpriv
-    sha256sum
     udevadm
     unshare
     useradd