--- a/tests/zfs-tests/include/libtest.shlib
+++ b/tests/zfs-tests/include/libtest.shlib
@@ -3455,33 +3455,24 @@ function tunable_exists
 }
 
 #
-# Compute MD5 digest for given file or stdin if no file given.
+# Compute xxh128sum for given file or stdin if no file given.
 # Note: file path must not contain spaces
 #
-function md5digest
+function xxh128digest
 {
-	openssl md5 -r $1 | awk '{print $1}'
+	xxh128sum $1 | awk '{print $1}'
 }
 
 #
-# Compare the MD5 digest of two files.
+# Compare the xxhash128 digest of two files.
 #
-function cmp_md5s {
+function cmp_xxh128 {
 	typeset file1=$1
 	typeset file2=$2
 
-	typeset sum1=$(md5digest $file1)
-	typeset sum2=$(md5digest $file2)
+	typeset sum1=$(xxh128digest $file1)
+	typeset sum2=$(xxh128digest $file2)
 	test "$sum1" = "$sum2"
-}
-
-#
-# Compute SHA256 digest for given file or stdin if no file given.
-# Note: file path must not contain spaces
-#
-function sha256digest
-{
-	openssl sha256 -r $1 | awk '{print $1}'
 }
 
 function new_fs #<args>