--- a/tests/zfs-tests/tests/functional/block_cloning/block_cloning.kshlib
+++ b/tests/zfs-tests/tests/functional/block_cloning/block_cloning.kshlib
@@ -28,8 +28,8 @@
 
 function have_same_content
 {
-	typeset hash1=$(md5digest $1)
-	typeset hash2=$(md5digest $2)
+	typeset hash1=$(xxh128digest $1)
+	typeset hash2=$(xxh128digest $2)
 
 	log_must [ "$hash1" = "$hash2" ]
 }