--- a/tests/zfs-tests/tests/functional/cli_root/zfs_receive/zfs_receive_corrective.ksh
+++ b/tests/zfs-tests/tests/functional/cli_root/zfs_receive/zfs_receive_corrective.ksh
@@ -72,7 +72,7 @@ function test_corrective_recv
 	log_must zpool status -v $TESTPOOL
 	log_mustnot eval "zpool status -v $TESTPOOL | \
 	    grep \"Permanent errors have been detected\""
-	typeset cksum=$(md5digest $file)
+	typeset cksum=$(xxh128digest $file)
 	[[ "$cksum" == "$checksum" ]] || \
 		log_fail "Checksums differ ($cksum != $checksum)"
 }
@@ -94,7 +94,7 @@ log_must zfs create -o primarycache=none \
 
 log_must dd if=/dev/urandom of=$file bs=1024 count=1024 oflag=sync
 log_must eval "echo 'aaaaaaaa' >> "$file
-typeset checksum=$(md5digest $file)
+typeset checksum=$(xxh128digest $file)
 
 log_must zfs snapshot $TESTPOOL/$TESTFS1@snap1
 
@@ -177,7 +177,7 @@ log_must zpool scrub -w $TESTPOOL
 log_must zpool status -v $TESTPOOL
 log_mustnot eval "zpool status -v $TESTPOOL | \
     grep \"Permanent errors have been detected\""
-typeset cksum=$(md5digest $file)
+typeset cksum=$(xxh128digest $file)
 [[ "$cksum" == "$checksum" ]] || \
 	log_fail "Checksums differ ($cksum != $checksum)"