--- a/tests/zfs-tests/tests/functional/cli_root/zpool_import/zpool_import.cfg
+++ b/tests/zfs-tests/tests/functional/cli_root/zpool_import/zpool_import.cfg
@@ -45,8 +45,8 @@ export MYTESTFILE=$STF_SUITE/include/libtest.shlib
 export CPATH=$TEST_BASE_DIR/cachefile.$$
 export CPATHBKP=$TEST_BASE_DIR/cachefile.$$.bkp
 export CPATHBKP2=$TEST_BASE_DIR/cachefile.$$.bkp2
-export MD5FILE=$TEST_BASE_DIR/md5sums.$$
-export MD5FILE2=$TEST_BASE_DIR/md5sums.$$.2
+export MD5FILE=$TEST_BASE_DIR/hashsums.$$
+export MD5FILE2=$TEST_BASE_DIR/hashsums.$$.2
 
 export GROUP_NUM=3
 typeset -i num=0