--- a/tests/zfs-tests/tests/functional/cp_files/cp_files_002_pos.ksh
+++ b/tests/zfs-tests/tests/functional/cp_files/cp_files_002_pos.ksh
@@ -61,8 +61,8 @@ function cleanup
 
 function verify_copy
 {
-	src_cksum=$(sha256digest $1)
-	dst_cksum=$(sha256digest $2)
+	src_cksum=$(xxh128digest $1)
+	dst_cksum=$(xxh128digest $2)
 
 	if [[ "$src_cksum" != "$dst_cksum" ]]; then
 		log_must ls -l $CP_TESTDIR