… tests/zfs-tests/tests/functional/rsend/send_encrypted_props.ksh