For data integrity checks as done in ZTS, the verification for unintended data corruption with `xxhash128` should be a lot faster and perfectly usable.

I removed the `sysutils/coreutils` dependency from FreeBSD and added `xxhash` everywhere.
The binaries `md5`, `md5sum`, `sha256` and `sha256sum` are also not needed anymore.

How faster a whole functional testrun will be isn't known. But we can compare this runs of this PR with older ones.

### Motivation and Context

Speedup ZTS a bit.

### Description


### How Has This Been Tested?

Timings for `2x zfs-tests.sh -vK -s 3GB -T rsend` are like this:

system | xxhash128 | md5 / sha256
-----------|-----------------|----------------------
qemu-x86 (almalinux8) | 26m 47s | 27m 50s
qemu-x86 (almalinux9) |28m 47s | 29m 55s
qemu-x86 (centos-stream9) | 29m 32s | 29m 53s
qemu-x86 (debian11) | 22m 56s | 25m 1s
qemu-x86 (debian12) | 24m 59s | 27m 44s
qemu-x86 (fedora39) | 29m 26s | 30m 50s
qemu-x86 (fedora40)|27m 19s | 29m 52s
qemu-x86 (freebsd13)|22m 43s | 21m 56s
qemu-x86 (freebsd13r)|21m 33s | 	21m 59s
qemu-x86 (freebsd14)|19m 55s | 21m 12s
qemu-x86 (freebsd14r)|20m 45s | 	23m 16s
qemu-x86 (ubuntu20)|22m 50s | 25m 5s
qemu-x86 (ubuntu22)|27m 38s | 29m 8s
qemu-x86 (ubuntu24)| 30m 51s | 	31m 52s
Setup + Cleanup | 32s | 31s
Total | 5h 56m 33s | 6h 16m 4s

Links:
- old: https://github.com/mcmilk/zfs/actions/runs/11075750256 - Total duration: 34m 17s
- new: https://github.com/mcmilk/zfs/actions/runs/11075382905 - Total duration 31m 51s

It's not much, but it is just the `rsend` test, which uses checksums a lot.

### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Performance enhancement (non-breaking change which improves efficiency)
- [x] Code cleanup (non-breaking change which makes code smaller or more readable)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Library ABI change (libzfs, libzfs\_core, libnvpair, libuutil and libzfsbootenv)
- [ ] Documentation (a change to man pages or other documentation)

### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the OpenZFS [code style requirements](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md#coding-conventions).
- [x] I have updated the documentation accordingly.
- [x] I have read the [**contributing** document](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md).
- [ ] I have added [tests](https://github.com/openzfs/zfs/tree/master/tests) to cover my changes.
- [x] I have run the ZFS Test Suite with this change applied.
- [x] All commit messages are properly formatted and contain [`Signed-off-by`](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md#signed-off-by).